rename: diary list 페이지의 컴포넌트를 features/diaryList로 이동
