+++ src/features/diaryList/apis/mutations.ts
@@ -1,12 +1,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { DIARY_KEY, docsApi } from '@/features/diary/diaryList/apis';
-import {
-  Doc,
-  DriveFile,
-  File,
-} from '@/features/diary/diaryList/apis/interfaces';
+import { DIARY_KEY, docsApi } from '@/features/diaryList/apis/index.ts';
+import { Doc, DriveFile, File } from '@/features/diaryList/apis/interfaces.ts';
 import axios, { AxiosError } from 'axios';
-import { DOMAIN_URI } from '@/features/diary/diaryList/apis';
+import { DOMAIN_URI } from '@/features/diaryList/apis/index.ts';
 
 const createDoc = async (title: string): Promise<Doc> =>
   await docsApi.post('', { title: `jiary-${title}` }).then(res => res.data);
