+++ src/features/diaryList/components/DiaryListHeader.tsx
@@ -1,7 +1,7 @@
-import * as style from '@/features/diary/diaryList/components/DiaryListHeader.css.ts';
+import * as style from '@/features/diaryList/components/DiaryListHeader.css';
 import Dropdown from '@/features/common/components/dropdown/Dropdown.tsx';
 import { useRef, useState } from 'react';
-import { useCreateDoc } from '@/features/diary/diaryList/apis/mutations';
+import { useCreateDoc } from '@/features/diaryList/apis/mutations.ts';
 import { useRouter } from 'next/router';
 
 export default function DiaryListHeader({ count }: { count: number }) {
